change fibonacci so it can return larger numbers, and remove other caches

+++ cache.h
@@ -13,8 +13,8 @@
 typedef int KeyType;
 #define KEY_FMT "%d"
 
-typedef int ValueType;
-#define VALUE_FMT "%d"
+typedef unsigned long long ValueType;
+#define VALUE_FMT "%llu"
