+++ fibonacci.c
@@ -2,10 +2,10 @@
 
 #include <stdio.h>
 
-int calculate_fibonacci(int nth) {
-    int first  = 0;
-    int second = 1;
-    int sum    = 0;
+unsigned long long calculate_fibonacci(int nth) {
+    unsigned long long first  = 0;
+    unsigned long long second = 1;
+    unsigned long long sum    = 0;
 
     for (int ix = 0; ix < nth; ix++) {
         sum    = first + second;
