+++ fibonacci.h
@@ -2,6 +2,6 @@
 #define FIBONACCI_H
 
 // calculate "nth" fibonacci number
-int calculate_fibonacci(int nth);
+unsigned long long calculate_fibonacci(int nth);
 
 #endif
